Remove rule about finishing ontop of the raised area
This is an incredibly difficult, functionally impossible task for teams to do anyway. But encouraging them to do this may lead to them attempting to ram into the arena, or focusing too much on trying to climb (and no doubt failing), and not enough time focusing on the rest of the game.